requirements/base.txt
django==4.2.6
djangorestframework==3.14.0
python-weixin==0.5.0
mysqlclient==2.2.0
PyJWT==2.8.0
django-cors-headers==4.2.0
django-redis==5.4.0
aliyun-python-sdk-core==2.14.0
drf-yasg==1.21.7
